Update README.md
"From now on..." implies that one will not use ad-supported Spotify anymore, which contradicts the declared intentions of this ad blocker. Certainly this slip-up does not reflect a deceptive declaration of intention. More likely, the wording "From now on..." was meant to apply to the scope of a party, meaning, "From now until the end of the party you are currently throwing." This pull request clarifies the author's declared intention of supplying this software in the good faith that nobody is going to use it to the exclusion of enjoying Spotify's excellent advertisements.